Beta version
Game play appears to be fully functional. Scoring has been added, scopa's are
processed correctly, and the game loops through until there is a winner. The
more obscure situations that may arise still need to be tested, particularly
with regard to multiple take options, but everything seems to be in good order.